--- lib/rust/ast/src/invalid.rs
@@ -1,15 +1,0 @@
-//! This module exports invalid ast shapes.
-
-
-
-// ===============
-// === Invalid ===
-// ===============
-
-
-// === Unrecognized ===
-
-/// Unrecognized token.
-#[allow(missing_docs)]
-#[derive(Debug,Clone)]
-pub struct Unrecognized { pub str: String }
