--- lib/rust/ast/src/number.rs
@@ -1,12 +1,0 @@
-//! This module exports ast shapes that represent numbers.
-
-
-
-// ==============
-// === Number ===
-// ==============
-
-/// The ast node for a number
-#[allow(missing_docs)]
-#[derive(Debug,Clone)]
-pub struct Number { pub number: String }
